not needed, as luasnip is working fine as a completion system outside of none-ls
For the sake of posterity:

I had trouble with this. It would error out when trying to load the
snippets through none-ls's completion system. Eventually, I tracked down
the problem to
https://github.com/nvimtools/none-ls.nvim/blob/37671797d6e5f9bd0ee3ab965ab2e6de251e1930/lua/null-ls/builtins/completion/luasnip.lua#L47
The `snippet.description` is a table, and nvim-cmp expects the detail to
be a string:
https://github.com/hrsh7th/nvim-cmp/blob/5260e5e8ecadaf13e6b82cf867a909f54e15fd07/lua/cmp/entry.lua#L449
It's hard to say which repo is in the wrong here. none-ls's line of code
predates luasnip's 1.0.0 version, so it's hard to know if none-ls's
completion ever worked. Maybe at some point, luasnip _did_ have some
sort of detail string, but I don't think we know what it was meant for.
I found that just commenting out the above none-ls line was effective at
resolving the bug.

I've asked a general question on the none-ls discussion board about it: https://github.com/nvimtools/none-ls.nvim/discussions/130 :shrug: